cosmetic changes - moved to post package - NavBar index fix - Create Post -> PostScreen
